Module rustc_dependencies::parse_format

Structs

Argument

Representation of an argument specification.

FormatSpec

Specification for the formatting of an argument in the format string.

InnerSpan

Range inside of a Span used for diagnostics when we only have access to relative positions.

InnerWidthMapping

The location and before/after width of a character whose width has changed from its source code representation

ParseError

Parser

The parser structure for interpreting the input format string. This is modeled as an iterator over Piece structures to form a stream of tokens being output.

Enums

Alignment

Enum of alignments which are supported.

Count

A count is used for the precision and width parameters of an integer, and can reference either an argument or a literal integer.

DebugHex

Enum for the debug hex flags.

ParseMode

The type of format string that we are parsing.

Piece

A piece is a portion of the format string which represents the next part to emit. These are emitted as a stream by the Parser class.

Position

Enum describing where an argument for a format can be located.

Sign

Enum for the sign flags.

Suggestion